Create or modify a style set that describes a full markdown text

A style set contains information on how to style the various tags in a markdown text. While it is not necessary to provide a style for all tags (it will just inherit the parent if missing), it is required to provide a complete style for the body tag so an option is avialable through inheritance for all tags and all style options. It can often be easier to derive a new style set from an existing one rather than building one from scratch.

style_set(...)

modify_style(style_set, tag, ...)

remove_style(style_set, tag)

Arguments

...

Named arguments providing a style for the specific tags. For modify_style() a number of style options to change. If the first argument is a marquee style it will overwrite the tag and subsequent arguments are ignored

style_set

A style set to modify

tag

The name of a tag to modify or remove

Value

A style set object

Examples

style_set(body = body_style(), p = style(indent = em(2))) |>
  modify_style("p", size = 16) |>
  modify_style("str", style(weight = "bold")) |>
  remove_style("p")
#> A marquee style set
#> body: size: 12, background: NA, family:, weight: 400, italic: FALSE, width: 5, features: <empty>, color: black, lineheight: 1.6, align: left, tracking: 0, indent: 0, hanging: 0, margin_top: 0, margin_right: 0, margin_bottom: rem(1), margin_left: 0, padding_top: 0, padding_right: 0, padding_bottom: 0, padding_left: 0, border: NA, border_size_top: 0, border_size_right: 0, border_size_bottom: 0, border_size_left: 0, border_radius: 0, bullets: •, ◦, ▪, ▫, ‣, ⁃, underline: FALSE, strikethrough: FALSE, img_asp: 1.65
#>  str: weight: 700